build fix, removed a hardwired debug option
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@4267 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/amr_mod.F90
@@ -9341,7 +9341,7 @@ endif
     use inputs_mod, ONLY : photoionPhysics, doselfgrav, extmass, sourcepos, sourceTeff, sourceRadius
     TYPE(octal), INTENT(INOUT) :: thisOctal
     INTEGER, INTENT(IN) :: subcell
-    real :: r, hr, rd
+    real(double) :: r, hr, rd
 !    real(double), parameter :: min_rho = 1.0d-24 ! minimum density
 !    real(double), parameter :: min_rho = 1.0d-21 ! minimum density
 !    real(double), parameter :: min_rho = 1.0d-22 ! minimum density
